# api/openapi_public.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TickerTrace - Institutional ETF Activity API",
    "description": "Real-time institutional ETF holdings data. Track daily position changes, cross-fund signals, sector flow, and option activity across ARK Invest, Avantis, Kurv, YieldMax, REX Shares, and more.\n\n**What you get:**\n- Daily holdings changes for 20+ actively-managed ETFs\n- Cross-fund divergence signals (same ticker, opposite directions)\n- Sector rotation flow analysis\n- Option writing activity tracking\n- Historical holdings data\n\n**Use cases:**\n- Quantitative research & alpha generation\n- Portfolio construction signals\n- Institutional flow monitoring\n- Alternative data feeds\n\n**Rate limits:**\n- Free: 100 requests/day\n- Pro ($15/mo): 5,000 requests/day\n- Institutional ($50/mo): 50,000 requests/day",
    "version": "1.1.0",
    "contact": {
      "name": "TickerTrace",
      "url": "https://ticker-trace.vercel.app",
      "email": "mphanko@gmail.com"
    },
    "x-logo": {
      "url": "https://ticker-trace.vercel.app/favicon.ico"
    }
  },
  "paths": {
    "/api/v1/signals": {
      "get": {
        "summary": "Get Signals",
        "description": "Full signal payload. Free tier \u2014 no API key needed.",
        "operationId": "get_signals_api_v1_signals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "Data"
        ]
      }
    },
    "/api/v1/stats": {
      "get": {
        "summary": "Get Stats",
        "description": "Global stats. Free tier.",
        "operationId": "get_stats_api_v1_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "Data"
        ]
      }
    },
    "/api/v1/sectors": {
      "get": {
        "summary": "Get Sectors",
        "description": "Sector flow. Free tier.",
        "operationId": "get_sectors_api_v1_sectors_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "Data"
        ]
      }
    },
    "/api/v1/changes": {
      "get": {
        "summary": "Get Changes",
        "description": "All daily position changes. Filterable by provider, fund, direction.",
        "operationId": "get_changes_api_v1_changes_get",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "name": "fund",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fund"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Direction"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Data"
        ]
      }
    },
    "/api/v1/fund/{fund}": {
      "get": {
        "summary": "Get Fund",
        "description": "Fund holdings detail \u2014 top holdings, options count, AUM.",
        "operationId": "get_fund_api_v1_fund__fund__get",
        "parameters": [
          {
            "name": "fund",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Fund"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Data"
        ]
      }
    },
    "/api/v1/ticker/{ticker}": {
      "get": {
        "summary": "Get Ticker",
        "description": "Cross-fund ticker detail \u2014 who's buying/selling this?",
        "operationId": "get_ticker_api_v1_ticker__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Data"
        ]
      }
    },
    "/api/v1/divergences": {
      "get": {
        "summary": "Get Divergences",
        "description": "Cross-fund divergences \u2014 same ticker, opposite directions.",
        "operationId": "get_divergences_api_v1_divergences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "Data"
        ]
      }
    },
    "/api/v1/funds": {
      "get": {
        "summary": "List Funds",
        "description": "List all tracked funds with provider and AUM.",
        "operationId": "list_funds_api_v1_funds_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "Data"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register",
        "description": "Register for a free API key with email + password.",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login",
        "description": "Log in with email + password. Returns API key for session auth.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get Me",
        "description": "Check your API key status, tier, and usage.",
        "operationId": "get_me_auth_me_get",
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/billing/checkout": {
      "post": {
        "summary": "Create Checkout",
        "description": "Create a Stripe Checkout session for a Pro or Institutional subscription.\nRedirects to Stripe-hosted checkout page.",
        "operationId": "create_checkout_billing_checkout_post",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          },
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "pro",
              "title": "Tier"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Billing"
        ]
      }
    },
    "/auth/redeem": {
      "post": {
        "summary": "Redeem Promo Code",
        "description": "Redeem a promo code to upgrade tier.\nUser must have an account (register first).",
        "operationId": "redeem_promo_code_auth_redeem_post",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "FirebaseLoginRequest": {
        "properties": {
          "id_token": {
            "type": "string",
            "title": "Id Token"
          }
        },
        "type": "object",
        "required": [
          "id_token"
        ],
        "title": "FirebaseLoginRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://api.tickertrace.mphinance.com",
      "description": "Production"
    }
  ]
}